+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [1.2.2] - 16-04-2026
+
+- Updated readme
+- Added latest workflows
+
+## [1.2.1] - 23-03-2026
+
+- Added unarchiving when a Person is archived and needs to be created again
+
+## [1.2.0] - 10-02-2026
+
+- Changed authentication method.
+- Added resources based on CostCenter
+- Added importscript
+- Added Enable and Disable
+- Updated readme
+
 ## [1.1.0] - 10-12-2024
 
 - Added option to rename resources.